Add Recipes pages
Background
==========
Add the recipes index page where a list of recipes is displayed and
every recipe card is a link to a recipe show page where more details
about the recipe is presented.

Technical Details
=================
The information is pulled from the Contentful's GraphQL API. Error
handling is added if a user tries to fetch some recipe that the API
doesn't store. This is done so we won't flood the logs with errors when
users try to tinker with the recipe ids.

Basic styling is done so the pages resemble the Marley Spoon's website.
Some rudimentary responsive design is done as well but this needs more
work if this site needs to be fully mobile friendly.

Additional Notes
================
I was wondering whether I should implement pagination for the index page
but since I don't know what the requirements are I decided to leave this
out for the moment until I have more clarity about the use cases for
this page.

I wanted to implement fetching the images in the appropriate sizes from
the API instead of fetching the default (probably the highest quality).
Apart from the reduced bandwidth this would give some confidence about
the image dimensions. Anyway, I decided to abandon this idea again due
to some missing information about the business requirements.